feat!: impl useGraphQlStyle hook
