+++ package.json
@@ -22,7 +22,7 @@
   "dependencies": {
     "@codemirror/lang-javascript": "^6.2.1",
     "@codemirror/lang-json": "^6.0.1",
-    "@codemirror/lint": "^6.4.2",
+    "@codemirror/lint": "^6.0.0",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
     "@fontsource/roboto": "^5.0.8",
@@ -33,6 +33,7 @@
     "@mui/material": "^5.15.1",
     "@uiw/codemirror-theme-bbedit": "^4.21.21",
     "@uiw/react-codemirror": "^4.21.21",
+    "cm6-graphql": "^0.0.12",
     "firebase": "^10.7.1",
     "graphql": "^16.8.1",
     "immer": "^10.0.3",
